USB: dwc3: msm: Don't disable/enable clocks if in lpm during block reset
Whenever interrupt is triggered for erratic event, USB driver schedules
work to do USB block reset as part of recovery. Currently while doing
block reset, not checking whether USB is in lpm and going ahead with
disabling clocks which messup clock counters. Hence check if USB is in lpm
and do disabling/enabling clocks only if it is not lpm.

Change-Id: Ib28e76f0786184506bec69b7541e56dd7a16ed0f
Signed-off-by: Vijayavardhan Vennapusa <vvreddy@codeaurora.org>