mdss: Only update the panel status after initializing
The livedisplay driver is using this flag to check whether the panel is
on to either perform work or ignore it.

However, the flag is being set to on before initialization completes and
causes glitches when turning on the screen while using livedisplay. This
patch moves the flag to the end of the initialization.

Change-Id: I4d67376a0f799537e5ed9979dbf061b5b4a3e13a